SBX-540. Refactor local template source creation and deletion
- Updated local template source creation to use new schema
- Added check to prevent deletion of the last local source
- Improved logging for source creation and deletion processes

+++ saltbox_core/task_templates/utils/orchestrator.py
@@ -629,6 +629,21 @@ class SyncOrchestrator:
         await self._source_service.update(
             source_id, {'current_operation': SourceOperation.REMOVE, 'current_task_id': task_id}
         )
+        is_last_local_source = await self._source_service.count(query={'source_type': SourceType.LOCAL_BUNDLE}) == 1
+
+        if is_last_local_source:
+            msg = 'Cannot remove the last local source. At least one local source must be present.'
+            logger.error(msg)
+            await self._source_service.update(
+                source_id,
+                {
+                    'state': SourceState.DISCOVERED,
+                    'current_operation': None,
+                    'last_error': msg,
+                    'current_task_id': None,
+                },
+            )
+            raise ValueError(msg)
         try:
             await self._source_service.delete(source_id)
         except Exception as e:
